requirements.txt
streamlit
python-dotenv
pdfplumber
pytesseract
pdf2image
Pillow
langchain
langchain-openai
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
google-generativeai>=0.7.2
gTTS
deep-translator
spacy
tiktoken
pymupdf
